Added Sample Dataset Cards to the Homepage
Modeled after the job-board homepage. The cards link directly to the page for the dataset. The "See More" button takes you to the Data catalog page. Also Changed the "Submit" button in the header to say "Contact Us", because it links to the "Contact Us" page.

Inefficieny: To display the sample datasets, I created a new list of datasets in the HomePage component because I wasn't sure how/if I should modify the BasicExample in index to somehow pass the dataset list as props to both the DataList and HomePage components.